refactor(capture): extract pure event filtering into query.ts
Scope/direction/since/limit semantics become probe-testable on plain data,
electron-free. No behavior change.
